Refacto ShoppingController : Move an rename Item and Boby class + moves Customertype in an enum
