+++ src/main/java/com/sipios/refactoring/model/CustomerType.java
@@ -1,0 +1,7 @@
+package com.sipios.refactoring.model;
+
+public enum CustomerType {
+    STANDARD_CUSTOMER,
+    PREMIUM_CUSTOMER,
+    PLATINUM_CUSTOMER
+}
