+++ src/main/java/com/sipios/refactoring/model/ItemQuantity.java
@@ -1,0 +1,21 @@
+package com.sipios.refactoring.model;
+
+public class ItemQuantity {
+
+    private final String type;
+    private final int nb;
+
+    public ItemQuantity(String type, int quantity) {
+        this.type = type;
+        this.nb = quantity;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public int getNb() {
+        return nb;
+    }
+
+}
